frsca-ml/src/provenance_generator/requirements.txt
cyclonedx-python-lib>=6.0.0
pydantic>=2.0.0
